pwm and adc fix
Found that when using the ADC to set the pwm for the discharge pin, the PWM that controls the DC converter would start to emit irregular pulses. Maybe it has something to do with DMA. Ultimately I found the solution was to remove setting the discharge PWM in the same function as reading the ADC. The Dc converter PWM now works fine. However it is not tested with power on so not sure if this is the last thing needed to get the converter injecting a sine wave into the load.