new pi working well, need to make clean function
the new pi loop uses a counter via interrupt to take an average reading that is greater than one EIS period, therefore the loop isn't trying to fight the sine wave of the EIS signal. Output is quite smooth on the power supply. Current floats +/- 20mA over a period of ~10s, but that's acceptable. Could be due to transients in current measurement or ADC issues.