can charge battery, but some issues
Using a dummy load, the algorithm successfully transitions from CC to CV. When connected to an actual battery, it works in either CC or CV modes, but starts slamming between the two when trying to transition. There also seems to be PI loop slamming when the I setpoint is too low.