channel 4 working. Channel 3 appears to have HW short
added ifdef lines to debug 2 channels separately. There is a HW short somewhere that crops up unexpectedly. I thought it might be a floating gate on the channel 3 discharge FET, but after observing for sure the gate is low, it still crops up. It may be a short somewhere caused by an earlier problem that I couldn't fully debug. So it's somewhere - maybe it'll be gone when I come back? >:( . All work on this branch has been focused on the L0 code. Will need to port changes over to F0 once two bats are working simultaneously.